+++ docs/CHANGELOG.md
@@ -7,6 +7,14 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.0.11] - 2025-10-22
+
+## Fixed
+
+- Connector start and stop methods wrapper
+
+- Freeze instances
+
 ## [1.0.10] - 2025-10-22
 
 ## Fixed